refactor(commands): extract /memory command into memory.rs
